requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Science & Computation
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Progress Bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Model Evaluation & Metrics
scikit-learn>=1.3.0

# Explainable AI (for XAI component)
captum>=0.6.0

# API / Web Service
flask>=3.0.0